[Yjs Collab] WIP - Sync changes on (auto)save using Yjs
Previously, Gutenberg implemented a basic Yjs editor binding. Changes
were synced using y-webrtc when collaborative editing is enabled.
However, the previous implementation did not reliably sync the
backend-state with the Yjs document. Hence, the previous implementation
could result in content duplication, or loss of changes that were
created in the collaborative session.

This commit implements the initial approach to sync the backend document
(HTML encoded) with the Yjs document without the server understanding
Yjs.

Following things were implemented:

- We save the Yjs document in an HTML comment, that is stored in the
  HTML document stored in WordPress
- When the client notices that the Yjs document stored in the comment
  does not reflect the state of the HTML document, we assume that the
  changes were generated by the backend. Hence we update the document
  that is stored in the HTML comment, and merge the Yjs update to the
  Yjs document that is used in the collaborative session. Then we save
  the merged document once again to the backend.

The changes from a backend are applied as a simulated "system user".
This should guarantee that everyone applies the changes in the same way,
so we avoid content duplication. The
approach in this commit is still insufficient. We need to store
additional information in the backend to ensure that certain edge-cases
are handled.