Add optional chain to sizes indexing of media details in edit-site
This optional chaining is already present in 21 other lines in Gutenberg where `sizes` (in `media_details`) is indexed, but is missing on these two lines.

For some reason, `media_details` is an empty object for some of my attachments on my site, so this throws a TypeError. This change resolves this issue.

Co-authored-by: grgar <grgarside@git.wordpress.org>