Simplify color specifications from `rgba` to `rgb`
Currently, the theme.json that ships in core at `wp-includes/theme.json` default gradients are inconsistent in color specification -- a number of places it uses `rgba(#,#,#,1)` even though that is equivalent to just `rgb(#,#,#)` -- there is no rhyme or reason I can see as to why rgba is used here, so this is to trim a couple bytes by serving them up as rgb instead of rgba.

As there is no transparency here, it may be slightly more efficient to use hex?  But I'm fine leaving it as RGB for less code churn in core unless anyone feels strongly here.