Filterable sync provider (#70539)
* [Yjs Collab] WIP - Sync changes on (auto)save using Yjs

Previously, Gutenberg implemented a basic Yjs editor binding. Changes
were synced using y-webrtc when collaborative editing is enabled.
However, the previous implementation did not reliably sync the
backend-state with the Yjs document. Hence, the previous implementation
could result in content duplication, or loss of changes that were
created in the collaborative session.

This commit implements the initial approach to sync the backend document
(HTML encoded) with the Yjs document without the server understanding
Yjs.

Following things were implemented:

- We save the Yjs document in an HTML comment, that is stored in the
  HTML document stored in WordPress
- When the client notices that the Yjs document stored in the comment
  does not reflect the state of the HTML document, we assume that the
  changes were generated by the backend. Hence we update the document
  that is stored in the HTML comment, and merge the Yjs update to the
  Yjs document that is used in the collaborative session. Then we save
  the merged document once again to the backend.

The changes from a backend are applied as a simulated "system user".
This should guarantee that everyone applies the changes in the same way,
so we avoid content duplication. The
approach in this commit is still insufficient. We need to store
additional information in the backend to ensure that certain edge-cases
are handled.

* [Yjs Collab] update used dependencies

* [Yjs Collab] fix collab issues introduced in upstream - ensure that entities are always registered.

* [Yjs Collab] Use heartbeat api to pull ydoc updates

* [Yjs Collab] cleanup

* [Yjs Collab] improve diffing

* [Yjs Collab] loss of selection & focus when content is saved

* [Yjs Collab] add origin to changes originating from gutenberg

* [Yjs Collab] documentation

* [Yjs Collab] improve docs

* [Yjs Collab] several bugfixes

* [Yjs Collab] add different feature flags for sync on save / heartbeat sync / webrtc sync

* [Yjs Collab] Fix typo in lib/experimental/synchronization.php

Co-authored-by: Aaron Jorbin <aaronjorbin@users.noreply.github.com>

* [Yjs Collab] user wp_rand instead of rand

Co-authored-by: Aaron Jorbin <aaronjorbin@users.noreply.github.com>

* [Yjs Collab] Update regexes and priorities for synchronization filters

As @aaronjorbin rightfully pointed out, the new filters approach don't
need to run at the end anymore.

* [Yjs Collab] implement get_yinfo function in synchronization.php

* [Yjs Collab] fix tests

* [Yjs Collab] fix php lint issues

* [Yjs Collab] fix mobile tests

* Remove heartbeat experiment

* Add core.getSyncProvider filter

---------

Co-authored-by: Kevin Jahns <kevin.jahns@protonmail.com>
Co-authored-by: Aaron Jorbin <aaronjorbin@users.noreply.github.com>