Force getBlockEditingMode to return 'contentOnly' or 'disabled' in isNavMode
We have a lot of logic spread throughout the codebase to determine if we're in contentOnly mode for a block. We're coupling it with isNavigationMode checks too. We already have this state within useBlockEditingMode, so it should simplify by only requiring the useBlockEditingMode check. This is assuming my understaning of how Write Mode is supposed to work is correct, which I'm not confident of yet :)