Avoid using `*-latest` for runner images
While using the `ubuntu-latest`, `macos-latest`, and `windows-latest` runner image tags is convenient, it has proven to be problematic in a number of instances as the runners are slowly updated (see Core-62808 and Core-62843).

Additionally, GitHub uses warnings on workflow runs using the `-latest` tags when there are upcoming changes to the images that the alias points to. There is no way to acknowledge and dismiss these notices, and they can be quite noisy and alarming.

This switches all workflows to using specific version tags representing the latest non-preview versions, which currently are as follows:

- `ubuntu-24.04`
- `windows-2025`
- `macos-15`