DataViews: preview size picker falls back to the smallest available size
When a grid dataview is passed a previewSize option which doesn't match
one of the discrete sizes available, it falls back to the next smallest
option. But if previewSize was smaller than the smallest available size
it would throw an exception.

The preview size picker control now uses the smallest available size if
the previewSize option is too small.